requirements.txt
requests==2.23.0
psycopg2_binary==2.8.5
colorama==0.4.3
beautifulsoup4==4.9.0
python_dateutil==2.8.1
PyYAML==5.3.1
